requirements.txt
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.6
boto3==1.9.135
botocore==1.12.135
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
confusable-homoglyphs==3.2.0
coverage==4.5.3
cryptography==2.6.1
Django==2.2
django-redis==4.10.0
django-registration==3.0
django-storages==1.7.1
djangorestframework==3.9.2
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.5
flake8==3.7.7
gunicorn==19.9.0
idna==2.7
isort==4.3.17
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
paramiko==2.4.2
pathtools==0.1.2
psycopg2-binary==2.8.2
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.1
PyNaCl==1.3.0
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
PyYAML==3.13
redis==3.2.1
requests==2.20.1
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
text-unidecode==1.2
texttable==0.9.1
typed-ast==1.3.4
urllib3==1.24.2
watchdog==0.9.0
websocket-client==0.56.0
wrapt==1.11.1
